fix(example-native): allow for cold start from Android Auto
